lxcu: Fix target host name on copy
lxc-copy does not change the target host name, which is odd.

The best way to change the host name would be to use systemd's
hostnamectl utility but this fails consistently in containers on
virtual machines, (this diagnostic has _not_ been exhaustively proven).

So, change the target host name using sed.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>