lxcu: Future proof OBS repo key
The problem is that the OBS repo key is set to expire. This is a problem
for APT systems.

The main part of the solution is the new package mgrant-obs-deb-keyring
which packages the keyring and can be updated before the key expires.

This, however, presents a chicken and egg situation; to install the key
package you must have a key. So, the 2 part solution is:-

1 ... At creation use curl to download the then current key. This then
  allows initial installs from the OBS repo.

2 ... As part of the normal basic/install.list install the
  mgrant-obs-deb-keyring package which then makes future key updates
  seemless.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>