at: Add notes why xmlstarlet not replaced by virt-xml
The replacement of xmlstarlet with virt-xml was investigated as it would
have been good to use a standard 'virt' interface. Unfortunately virt-xml
does not interrogate all aspects of the xml and so grep or xmlstarlet
must be used. (An example is determining whether a snapshot is internal
or external, see lqvm-snap.sh.in).

An external example showing grep being used can be seen at:-
Link: https://fabianlee.org/2021/01/10/kvm-creating-and-reverting-libvirt-external-snapshots/

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>