lxcu: bash: Set locale without using localectl in Debian
On one machine creating an lxc debian trixie container from a bookworm
host works, while on another it fails. The failure results in a dbus
message saying Access denied. It is not clear at all what the problem is.
So avoid the situation be replacing a localectl call with manually
setting the locale.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>